+++ frontend/app/note/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { PageHeader } from "@/lib/components/PageHeader/PageHeader";
-import { TextEditor } from "@/lib/components/TextEditor";
+import { TextEditor } from "@/lib/components/TextEditor/TextEditor";
 
 import styles from "./page.module.scss";
