--- frontend/lib/components/TextEditor/components/Toolbar/index.ts
@@ -1,1 +1,0 @@
-export * from "./Toolbar"
