--- frontend/lib/components/TextEditor/components/ToolbarButton/index.ts
@@ -1,1 +1,0 @@
-export * from "./ToolbarButton";
