--- frontend/lib/components/TextEditor/components/index.ts
@@ -1,2 +1,0 @@
-export * from "./ToolbarButton";
-export * from "./Toolbar";
